ext2spice/ext2spice.c: Multiplication result converted to larger type
Fix code scanning alert no. 47: Multiplication result converted to larger type (#62)

+++ ext2spice/ext2spice.c
@@ -2184,7 +2184,7 @@ spcWriteParams(dev, hierName, scale, l, w, sdM)
 		if (esScale < 0)
 		    fprintf(esSpiceF, "%g", w * scale);
 		else if (plist->parm_scale != 1.0)
-		    fprintf(esSpiceF, "%g", w * scale * esScale
+		    fprintf(esSpiceF, "%g", (double)w * scale * esScale
 				* plist->parm_scale * 1E-6);
 		else
 		    esSIvalue(esSpiceF, 1.0E-6 * (w + plist->parm_offset)
